:sparkles: useThingAttributeSystem utilizes context
